Fixed Search Icon Not Centering in Label Element
Justify self and Align self only work on the first direct child of a grid.
Since label is the new direct child element of the grid, Js and As only work on label, not the icon.
Changed the label width and height to match the icon, and used Js and As to align.